controllers,internal: Move the logic for determining failing POs in the cluster to the util package
